[refactor] dagster._core.definitions.solid_container -> node_container
